requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx>=0.27.0
transformers>=4.47.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
accelerate>=0.20.0
qwen-vl-utils>=0.0.8
